requirements.txt
django==5.2.7
djangorestframework==3.15.*
pandas==2.2.*
openpyxl==3.1.*
python-dotenv==1.0.*
psycopg2-binary==2.9.*
